+++ whistle-on-backend/src/main/java/com/halaguys/whistleon/dto/response/TeamResponseDto.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Getter
 @NoArgsConstructor
 public class TeamResponseDto {
-    private Long teamId;
+    private int teamId;
     private String teamName;
     private String location;
     private String logo;
@@ -18,7 +18,7 @@ public class TeamResponseDto {
     private List<User> users;
 
     @Builder
-    public TeamResponseDto(Long teamId, String teamName, String location, String logo, String description, List<User> users) {
+    public TeamResponseDto(int teamId, String teamName, String location, String logo, String description, List<User> users) {
         this.teamId = teamId;
         this.teamName = teamName;
         this.location = location;
